fix concurrency

+++ localstack/services/sqs/provider.py
@@ -541,22 +541,15 @@ class SqsQueue:
 
             return copied_message
 
-    def purge(self):
+    def clear(self):
+        """
+        Calls clear on all internal datastructures that hold messages and data related to them.
+        """
         with self.mutex:
-            if self.purge_in_progress:
-                raise PurgeQueueInProgress(
-                    f"Only one PurgeQueue operation on {self.name} is allowed every 60 seconds."
-                )
-
-            self.purge_timestamp = time.time()
-            self.purge_in_progress = True
-            try:
-                self.visible.queue.clear()
-                self.inflight.clear()
-                self.delayed.clear()
-                self.receipts.clear()
-            finally:
-                self.purge_in_progress = False
+            self.visible.queue.clear()
+            self.inflight.clear()
+            self.delayed.clear()
+            self.receipts.clear()
 
     def create_receipt_handle(self, message: SqsMessage) -> str:
         return encode_receipt_handle(self.arn, message)
@@ -1408,13 +1401,14 @@ class SqsProvider(SqsApi, ServiceLifecycleHook):
     def purge_queue(self, context: RequestContext, queue_url: String) -> None:
         queue = self._resolve_queue(context, queue_url=queue_url)
 
-        if config.SQS_DELAY_PURGE_RETRY and queue.purge_timestamp:
-            if (queue.purge_timestamp + 60) > time.time():
-                raise PurgeQueueInProgress(
-                    f"Only one PurgeQueue operation on {queue.name} is allowed every 60 seconds."
-                )
-
-        queue.purge()
+        with queue.mutex:
+            if config.SQS_DELAY_PURGE_RETRY:
+                if queue.purge_timestamp and (queue.purge_timestamp + 60) > time.time():
+                    raise PurgeQueueInProgress(
+                        f"Only one PurgeQueue operation on {queue.name} is allowed every 60 seconds."
+                    )
+            queue.purge_timestamp = time.time()
+            queue.clear()
 
     def set_queue_attributes(
         self, context: RequestContext, queue_url: String, attributes: QueueAttributeMap
